Sample Data of Orders on the Deliveraide Mobile Application

A tibble containing 20 sample orders in the format of data exported from the mobile application. The data can be used for testing and running of examples. The variables are as follows:

Usage

sample_orders_data

Format

A tibble with 20 rows and 17 variables:

city

Name of the city where the order is placed

zip

ZIP code of the city where the order is placed

meat

Indicator of whether the order includes meat (1 = includes meat, 0 = does not include meat)

fish

Indicator of whether the order includes fish (1 = includes fish, 0 = does not include fish)

porc

Indicator of whether the order includes pork (1 = includes pork, 0 = does not include pork)

is_completed

Indicator of whether the order is completed (1 = completed, 0 = not completed)

tracking_status

Status of order tracking (0 = not yet accepted, 1 = waiting for time slot, 2 = time slot convened with courier, 3 = delivered)

createdAt

Date and time of when the order was created

updatedAt

Date and time of when the order was last updated (if tracking_status == 3, then date and time of kit delivered)

date_accepted

Date and time of when the order was accepted

date_request

Date and time of when the order request was made

priority_score

Priority score assigned to the beneficiary

is_alone

Indicator of whether the beneficiary lives alone (1 = alone, 0 = not alone)

is_apl

Indicator of whether the beneficiary earns APL (1 = APL, 0 = not APL)

is_boursier

Indicator of whether the beneficiary is a scholarship recipient (1 = boursier, 0 = not boursier)

is_handicap

Indicator of whether the beneficiary has a disability (1 = disability, 0 = not disability)

is_winning

Indicator of whether the beneficiary is in the active workforce (1 = winning, 0 = not winning)